+++ tests/core/test_tab.py
@@ -164,3 +164,45 @@ async def test_wait_for_ready_state(browser: zd.Browser):
 
     ready_state = await tab.evaluate("document.readyState")
     assert ready_state == "complete"
+
+
+async def test_expect_request(browser: zd.Browser):
+    tab = browser.main_tab
+
+    async with tab.expect_request(sample_file("groceries.html")) as request_info:
+        await tab.get(sample_file("groceries.html"))
+        req = await asyncio.wait_for(request_info.value, timeout=3)
+        assert type(req) is zd.cdp.network.RequestWillBeSent
+        assert type(req.request) is zd.cdp.network.Request
+        assert req.request.url == sample_file("groceries.html")
+        assert req.request_id is not None
+
+        response_body = await request_info.response_body
+        assert response_body is not None
+        assert type(response_body) is tuple
+
+
+async def test_expect_response(browser: zd.Browser):
+    tab = browser.main_tab
+
+    async with tab.expect_response(sample_file("groceries.html")) as response_info:
+        await tab.get(sample_file("groceries.html"))
+        resp = await asyncio.wait_for(response_info.value, timeout=3)
+        assert type(resp) is zd.cdp.network.ResponseReceived
+        assert type(resp.response) is zd.cdp.network.Response
+        assert resp.request_id is not None
+
+        response_body = await response_info.response_body
+        assert response_body is not None
+        assert type(response_body) is tuple
+
+
+async def test_expect_download(browser: zd.Browser):
+    tab = browser.main_tab
+
+    async with tab.expect_download() as download_ex:
+        await tab.get(sample_file("groceries.html"))
+        await (await tab.select("#download_file")).click()
+        download = await asyncio.wait_for(download_ex.value, timeout=3)
+        assert type(download) is zd.cdp.browser.DownloadWillBegin
+        assert download.url is not None